+++ src/CleanRegex/Internal/Needle.php
@@ -1,0 +1,40 @@
+<?php
+namespace TRegx\CleanRegex\Internal;
+
+use TRegx\CleanRegex\Internal\Expression\Predefinition\Predefinition;
+use TRegx\CleanRegex\Internal\Split\ChainLinks;
+
+class Needle
+{
+    /** @var Predefinition */
+    private $predefinition;
+
+    public function __construct(Predefinition $predefinition)
+    {
+        $this->predefinition = $predefinition;
+    }
+
+    public function twoPieces(string $subject): array
+    {
+        $cut = new Cut($this->predefinition->definition());
+        return $cut->twoPieces($subject);
+    }
+
+    public function splitAll(string $subject): array
+    {
+        $links = new ChainLinks($this->predefinition->definition());
+        return $links->links($subject);
+    }
+
+    public function splitFromStart(string $subject, int $maxSplits): array
+    {
+        $links = new ChainLinks($this->predefinition->definition());
+        return $links->linksFromStart($subject, $maxSplits);
+    }
+
+    public function splitFromEnd(string $subject, int $maxSplits): array
+    {
+        $links = new ChainLinks($this->predefinition->definition());
+        return $links->linksFromEnd($subject, $maxSplits);
+    }
+}