+++ test/Feature/CleanRegex/PatternTest.php
@@ -163,30 +163,6 @@ class PatternTest extends TestCase
     /**
      * @test
      */
-    public function shouldGet_split()
-    {
-        // when
-        $matches = pattern(',')->split('Foo,Bar,Cat');
-
-        // then
-        $this->assertSame(['Foo', 'Bar', 'Cat'], $matches);
-    }
-
-    /**
-     * @test
-     */
-    public function shouldGet_split_WithCapturingGroup()
-    {
-        // when
-        $matches = pattern('(,)')->split('Foo,Bar,Cat');
-
-        // then
-        $this->assertSame(['Foo', ',', 'Bar', ',', 'Cat'], $matches);
-    }
-
-    /**
-     * @test
-     */
     public function shouldGet_literal()
     {
         // when